+++ packages/ui/src/components/SearchField/index.ts
@@ -1,0 +1,2 @@
+export * from "./SearchField";
+export type * from "./types";